Organize merged files
